backend/database/requirements.txt
supabase == 2.9.1
fastapi 
uvicorn
python-dotenv
httpx
websockets